+++ src/modules/data/DataProvider.tsx
@@ -34,6 +34,7 @@ import * as materialRequestMutations from './mutations/materialRequestMutations'
 import * as purchaseRequestMutations from './mutations/purchaseRequestMutations';
 import * as genericMutations from './mutations/genericMutations';
 import * as paymentStateMutations from './mutations/paymentStateMutations';
+import * as budgetMutations from './mutations/budgetMutations';
 import * as rfqMutations from './mutations/rfqMutations';
 import * as receptionMutations from './mutations/receptionMutations';
 import * as costCenterMutations from './mutations/costCenterMutations';
@@ -540,6 +541,7 @@ function useAppValue(): readonly [AppStateContextType, React.Dispatch<AppStateAc
         approvePaymentState: bindContext(paymentStateMutations.approvePaymentState),
         markPaymentStatePaid: bindContext(paymentStateMutations.markPaymentStatePaid),
         annulPaymentState: bindContext(paymentStateMutations.annulPaymentState),
+        addBudgetEntry: bindContext(budgetMutations.addBudgetEntry),
 
         // Safety
         addChecklistTemplate: bindContext(safetyMutations.addChecklistTemplate),